embedding/requirements.txt
# Requirements for the embedding pipeline
# Core OCR and document processing
PyPDF2>=3.0.1
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0
PyMuPDF>=1.23.0

# TrOCR and transformers
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0

# Embedding and vector database
sentence-transformers>=2.2.2
qdrant-client>=1.6.0

# SharePoint integration
msal>=1.20.0
requests>=2.31.0

# Utilities
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.65.0
